Allow for toggling of layout hierarchy compression
When the compression feature is set, the accessibility
layout hierarchy will only show important nodes and will
remove unnecessary surrounding layouts making the hierarchy
more human readable and efficient to search.

The compressed hierarchy will be the default. Previous tests
which depended on the more verbose hierarchy, will either
need to update their selectors, or disable this feature
from the UiDevice.

However, the uiautomatorviewer will always show the hierarchy
in its compressed form to encourage the user of the more
compact and efficient selectors.

Change-Id: I1d5d909ce7d4f60108dc491348fcf8da8549806c