adding IAutomationSupport for API compatibility
We didn't get IAutomationSupport in first pass of instrumentation
based library. This was because IAutomationSupport was originally
used in shell based library to mimic the behavior of a real
instrumentation, specially the sendStatus function. So it seemed
redundant.

However, for full API compatibility (of instrumentation library)
against shell library, we should have this. And it's just a thin
wrapper around an Instrumentation class to provide sendStatus
functionality.

Change-Id: I307fe2932653f7b9993dab6a54d46da2b3ba8eb3